% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm33_to_leaflet.R
\name{utm33_to_leaflet}
\alias{utm33_to_leaflet}
\title{Convert map from UTM 33 projection to epsg:4326 projection}
\usage{
utm33_to_leaflet(map, sf = TRUE)
}
\arguments{
\item{map}{Map to be converted}

\item{sf}{If true: the map is of format sf. If false: the map is of format sp}
}
\value{
Converted map
}
\description{
Convert map from UTM 33 projection to epsg:4326 projection
}
